add gumdrop, add --begin-step opt

+++ nlp_controller.py
@@ -14,6 +14,7 @@ from nlp_modules.marmot_tagger import MarmotTagger
 from nlp_modules.snlp_parser import StanfordNLPParser
 from nlp_modules.tt_tagger import TreeTaggerTagger
 from nlp_modules.tt_tokenizer import TreeTaggerTokenizer
+from nlp_modules.gumdrop_splitter import GumdropSplitter
 
 SCRIPT_DIR = os.path.dirname(os.path.realpath(__file__)) + os.sep
 LIB_DIR = SCRIPT_DIR + "lib" + os.sep
@@ -25,7 +26,13 @@ MODULES = {
     "tt_tagger": TreeTaggerTagger,
     "marmot_tagger": MarmotTagger,
     "snlp_parser": StanfordNLPParser,
+    "gumdrop_splitter": GumdropSplitter,
 }
+
+
+def lazy_pipeline(slugs, opts):
+    for slug in slugs:
+        yield MODULES[slug](opts)
 
 
 class NLPController:
@@ -73,11 +80,11 @@ class NLPController:
                 f"Checking dependencies for module {module.__class__.__name__}..."
             )
             module.test_dependencies()
-        # TODO: add dependencies information to each module and verify that this is a valid pipeline
 
         logging.info("NLPController initialization complete.\n")
 
     def _init_output_dir(self, initial_dir_path):
+        """Copy the input directory into the output directory."""
         os.makedirs(os.path.join(initial_dir_path, "xml"), exist_ok=True)
 
         filepaths = glob(os.path.join(self.input_dir, "**/*.xml"), recursive=True)
@@ -90,9 +97,29 @@ class NLPController:
         logging.info(f"Done copying initial files.\n")
 
     def run(self):
-        last_dir_name = os.path.join(self.output_dir, "00_initial")
-        self._init_output_dir(last_dir_name)
-        for i, module in enumerate(self.pipeline):
+        """Create the output directory and run every step of the pipeline in sequence,
+        creating a fresh directory for each step."""
+        begin_step = self.opts["begin_step"]
+
+        # init dirs if we're not skipping a step
+        if begin_step == 0:
+            last_dir_name = os.path.join(self.output_dir, "00_initial")
+            self._init_output_dir(last_dir_name)
+        # if we are skipping steps, delete the dirs after the skipped steps
+        else:
+            last_dir_name = glob(
+                os.path.join(self.output_dir, str(begin_step).zfill(2) + "*")
+            )[0]
+            dirs_to_delete = [
+                glob(os.path.join(self.output_dir, str(i).zfill(2) + "*"))[0]
+                for i in range(begin_step + 1, len(self.pipeline) + 1)
+            ]
+            for dirname in dirs_to_delete:
+                print(f"removing {dirname}")
+                shutil.rmtree(dirname)
+
+        steps = list(enumerate(self.pipeline))[self.opts["begin_step"] :]
+        for i, module in steps:
             input_dir = last_dir_name
             output_dir = os.path.join(
                 self.output_dir, str(i + 1).zfill(2) + "_" + module.__class__.__name__
@@ -104,130 +131,6 @@ class NLPController:
             last_dir_name = output_dir
 
 
-#        # the files we need to process
-#        INPUT_FILE_DIR = SCRIPT_DIR + "out" + os.sep + genre + os.sep + "autogum_*.xml"
-#        filepaths = sorted(glob(INPUT_FILE_DIR))
-#
-#        for file_num, filepath in enumerate(filepaths):
-#            with open(filepath, "r", encoding="utf8") as f:
-#                raw_xml = f.read()
-#            tokenized = tokenize(raw_xml)
-#
-#            tok_count = len(
-#                [t for t in tokenized.strip().split("\n") if not t.startswith("<")]
-#            )
-#
-#            sys.stderr.write(
-#                "o Processing document: "
-#                + os.path.basename(filepath)
-#                + " ("
-#                + str(file_num + 1)
-#                + "/"
-#                + str(len(filepaths))
-#                + ")"
-#            )
-#
-#            # Skip documents that are way too big or small
-#            if tok_count < 300 or tok_count > 2000:
-#                sys.stderr.write(" [skipped due to size]\n")
-#                continue
-#            else:
-#                sys.stderr.write("\n")
-#
-#            # POS tag
-#            # If we want to tag outside StanfordNLP, a dedicated tagger can be used
-#            if opts.no_parse:
-#                tagged = pos_tag(tokenized)
-#
-#            # Add sentence splits - note this currently produces mal-nested SGML
-#            if opts.no_sent:
-#                split_indices = [1] + [0] * (len(tokenized) - 1)
-#            else:
-#                from lib.gumdrop.EnsembleSentencer import EnsembleSentencer
-#
-#                best_sentencer_ever = EnsembleSentencer(
-#                    lang="eng", model="eng.rst.gum", genre_pat="_([^_]+)_"
-#                )
-#                split_indices = best_sentencer_ever.predict(
-#                    tokenized, as_text=True, plain=True, genre=genre
-#                )
-#
-#            counter = 0
-#            splitted = []
-#            opened_sent = False
-#            para = True
-#            for line in tokenized.strip().split("\n"):
-#                if not (line.startswith("<") and line.endswith(">")):
-#                    # Token
-#                    if split_indices[counter] == 1 or para:
-#                        if opened_sent:
-#                            splitted.append("</s>")
-#                            opened_sent = False
-#                        splitted.append("<s>")
-#                        opened_sent = True
-#                        para = False
-#                    counter += 1
-#                elif (
-#                        "<p>" in line or "<head>" in line or "<caption>" in line
-#                ):  # New block, force sentence split
-#                    para = True
-#                splitted.append(line)
-#            splitted = "\n".join(splitted)
-#            if opened_sent:
-#                if splitted.endswith("</text>"):
-#                    splitted = splitted.replace("</text>", "</s>\n</text>")
-#                else:
-#                    splitted += "\n</s>"
-#
-#            if not opts.no_parse:
-#                # Parse
-#                no_xml = splitted.replace("</s>\n<s>", "---SENT---")
-#                no_xml = re.sub(r"<[^<>]+>\n?", "", no_xml)
-#
-#                sents = no_xml.strip().replace("\n", " ").replace("---SENT--- ", "\n")
-#                parsed = dep_parse(sents, snlp, torch)
-#            else:
-#                parsed = tagged
-#
-#            doc = os.path.basename(filepath)
-#
-#            # Insert tags into XML
-#            pos_lines = []
-#            lemma_lines = []
-#            for line in parsed.split("\n"):
-#                if "\t" in line:
-#                    fields = line.split("\t")
-#                    if opts.no_parse:
-#                        lemma, xpos = fields[2], fields[1]
-#                    else:
-#                        lemma, xpos = fields[2], fields[4]
-#                    pos_lines.append(xpos)
-#                    lemma_lines.append(lemma)
-#            tagged = []
-#            counter = 0
-#            for line in splitted.split("\n"):
-#                if line.startswith("<") and line.endswith(">"):
-#                    tagged.append(line)
-#                else:
-#                    line = line + "\t" + pos_lines[counter] + "\t" + lemma_lines[counter]
-#                    tagged.append(line)
-#                    counter += 1
-#            tagged = "\n".join(tagged)
-#
-#            # Write output files
-#            with io.open(XML_OUTPUT_DIR + doc, "w", encoding="utf8", newline="\n") as f:
-#                f.write(tagged)
-#
-#            if not opts.no_parse:
-#                with io.open(
-#                        DEPENDENCY_OUTPUT_DIR + doc.replace(".xml", ".conllu"),
-#                        "w",
-#                        encoding="utf8",
-#                        newline="\n",
-#                ) as f:
-#                    f.write(parsed)
-
-
 def main():
     p = ArgumentParser()
     p.add_argument("output_dir", help="The directory that output should be written to.")
@@ -236,7 +139,7 @@ def main():
         "--modules",
         nargs="+",
         choices=MODULES.keys(),
-        default=["tt_tokenizer", "tt_tagger"],
+        default=["tt_tokenizer", "tt_tagger", "gumdrop_splitter"],
         help="NLP pipeline modules, included in the order they are specified.",
     )
     p.add_argument(
@@ -259,21 +162,33 @@ def main():
         action="store_true",
         help="Modules will attempt to use GPU if this flag is provided.",
     )
+    p.add_argument(
+        "--begin-step",
+        type=int,
+        default=0,
+        help=(
+            "If provided, will begin the pipeline from the ZERO-INDEXED step "
+            "in the pipeline corresponding to this value. E.g., --resume-step 1"
+            " would resume step y in the pipeline [x y z]. Every other step "
+            "before the resumed step is assumed to have been executed successfully."
+            " If the value if this parameter is greater than 0, --overwrite is ignored."
+        ),
+    )
     opts = p.parse_args()
 
     # Check if output directory already exists.
     if os.path.exists(opts.output_dir):
-        if opts.overwrite:
+        if opts.overwrite and opts.begin_step == 0:
             logging.warning(
                 f"About to delete ALL data from {opts.output_dir}. Interrupt now if you want to keep it."
             )
-            for i in range(5, 0, -1):
+            for i in range(3, 0, -1):
                 print(f"Deleting in {i}s...\r", end="")
                 sleep(1)
             shutil.rmtree(opts.output_dir)
             os.mkdir(opts.output_dir)
             logging.info(f"Deleted and re-created {opts.output_dir}.\n")
-        else:
+        elif opts.begin_step == 0:
             raise Exception(
                 "Output path "
                 + opts.output_dir
